+++ src/main/java/com/pccore/pccore/model/Ventas.java
@@ -1,19 +1,6 @@
 package com.pccore.pccore.model;
 
-import java.time.OffsetDateTime;
-import java.util.List;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "ventas")
@@ -25,7 +12,7 @@ public class Ventas {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private double total;
+    private BigDecimal total;
 
     @Column(insertable = true, updatable = false)
     private OffsetDateTime fecha;
